Modify testcase for validate update renewal threshold.
